% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{facets_theme}
\alias{facets_theme}
\title{Facets plotting theme}
\usage{
facets_theme(base_size = 11, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Base text size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Facets plotting theme
}
