% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the distribution of measures for a facet}
\usage{
plot_distribution(x, facet_index = 2L, binwidth = 0.75)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}

\item{facet_index}{Which measurement report to plot (default `2L`).}

\item{binwidth}{Histogram bin width in logits.}
}
\value{
A ggplot object.
}
\description{
Plot the distribution of measures for a facet
}
