% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot measure estimates with standard errors for a facet}
\usage{
plot_estimates(x, facet_index = 3L)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}

\item{facet_index}{Which measurement report to plot (default `3L`).}
}
\value{
A ggplot object.
}
\description{
Plot measure estimates with standard errors for a facet
}
