% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fairness}
\alias{plot_fairness}
\title{Plot representative observed versus fair averages for a facet}
\usage{
plot_fairness(x, facet_index = 1L, n = 8L)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}

\item{facet_index}{Which measurement report to plot (default `1L`).}

\item{n}{Number of representative elements to show.}
}
\value{
A ggplot object.
}
\description{
Plot representative observed versus fair averages for a facet
}
