% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_structure}
\alias{plot_structure}
\title{Plot mode, median, and mean category regions from rating-scale thresholds}
\usage{
plot_structure(x, theta = NULL)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}

\item{theta}{Optional vector of theta values.}
}
\value{
A ggplot object.
}
\description{
Plot mode, median, and mean category regions from rating-scale thresholds
}
