% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_unexpected}
\alias{plot_unexpected}
\title{Plot the strongest unexpected responses from Table 4.1}
\usage{
plot_unexpected(x, top_n = 20L)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}

\item{top_n}{Number of rows to show.}
}
\value{
A ggplot object.
}
\description{
Plot the strongest unexpected responses from Table 4.1
}
