% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_usage}
\alias{plot_usage}
\title{Plot rating-scale category usage frequencies}
\usage{
plot_usage(x)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}
}
\value{
A ggplot object.
}
\description{
Plot rating-scale category usage frequencies
}
