% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_wright}
\alias{plot_wright}
\title{Plot a compact Wright map across available facets}
\usage{
plot_wright(x)
}
\arguments{
\item{x}{A `facets_out` object or path to a FACETS `.out` file.}
}
\value{
A ggplot object.
}
\description{
Plot a compact Wright map across available facets
}
