% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_app}
\alias{run_app}
\title{Launch the facetsviz Shiny app}
\usage{
run_app(launch = TRUE, host = "127.0.0.1", port = NULL)
}
\arguments{
\item{launch}{If `TRUE`, run the app immediately. If `FALSE`, return the app object.}

\item{host}{Host passed to [shiny::runApp()] when launching.}

\item{port}{Port passed to [shiny::runApp()] when launching.}
}
\value{
A shiny app object when `launch = FALSE`, otherwise the result of [shiny::runApp()].
}
\description{
Launch the facetsviz Shiny app
}
